# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Moonshit Architect API",
    "version": "5.0.0",
    "description": "Full-stack AI DevOps API enabling GPT to edit the website, upload assets, run Python, commit files to GitHub, restore history, trigger builds, download backups, and inspect logs."
  },
  "servers": [
    {
      "url": "https://gpt-site-live-78f2a81c6107.herokuapp.com",
      "description": "Production Server"
    }
  ],
  "paths": {
    "/update-site": {
      "post": {
        "summary": "Create or update a text file in /public",
        "operationId": "updateSiteFile",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "filename": { "type": "string" },
                  "content": { "type": "string" }
                },
                "required": ["filename", "content"]
              }
            }
          }
        },
        "responses": { "200": { "description": "File written successfully." } }
      }
    },

    "/read-file": {
      "post": {
        "summary": "Read a file from /public",
        "operationId": "readFile",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "filename": { "type": "string" }
                },
                "required": ["filename"]
              }
            }
          }
        },
        "responses": { "200": { "description": "File content returned." } }
      }
    },

    "/delete-file": {
      "post": {
        "summary": "Delete a file in /public",
        "operationId": "deleteFile",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "filename": { "type": "string" }
                },
                "required": ["filename"]
              }
            }
          }
        },
        "responses": { "200": { "description": "File deleted." } }
      }
    },

    "/list-files": {
      "get": {
        "summary": "List all files in /public recursively",
        "operationId": "listFiles",
        "responses": { "200": { "description": "Files listed." } }
      }
    },

    "/upload-asset": {
      "post": {
        "summary": "Upload binary asset to /public using base64",
        "operationId": "uploadAsset",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "filename": { "type": "string" },
                  "base64": { "type": "string" }
                },
                "required": ["filename", "base64"]
              }
            }
          }
        },
        "responses": { "200": { "description": "Asset uploaded." } }
      }
    },

    "/trigger-build": {
      "post": {
        "summary": "Trigger the site build pipeline",
        "operationId": "triggerBuild",
        "responses": { "200": { "description": "Build triggered." } }
      }
    },

    "/run-python": {
      "post": {
        "summary": "Execute a Python script inside the sandbox",
        "operationId": "runPython",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "filename": { "type": "string" },
                  "code": { "type": "string" }
                },
                "required": ["filename", "code"]
              }
            }
          }
        },
        "responses": { "200": { "description": "Python executed." } }
      }
    },

    "/commit-file": {
      "post": {
        "summary": "Commit a single file to the GitHub repository",
        "operationId": "commitFile",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "filename": { "type": "string" },
                  "content": { "type": "string" },
                  "commit_message": { "type": "string" }
                },
                "required": ["filename", "content", "commit_message"]
              }
            }
          }
        },
        "responses": { "200": { "description": "File committed to GitHub." } }
      }
    },

    "/commit-changes": {
      "post": {
        "summary": "Commit all current changes to the GitHub repository",
        "operationId": "commitChanges",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "commit_message": { "type": "string" }
                },
                "required": ["commit_message"]
              }
            }
          }
        },
        "responses": { "200": { "description": "Changes committed to GitHub." } }
      }
    },

    "/list-commits": {
      "get": {
        "summary": "View recent GitHub commit history",
        "operationId": "listCommits",
        "responses": { "200": { "description": "Commit history returned." } }
      }
    },

    "/restore-commit": {
      "post": {
        "summary": "Restore a file from a given commit hash",
        "operationId": "restoreCommit",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "commit_hash": { "type": "string" },
                  "filename": { "type": "string" }
                },
                "required": ["commit_hash", "filename"]
              }
            }
          }
        },
        "responses": { "200": { "description": "File restored from history." } }
      }
    },

    "/backup-site": {
      "get": {
        "summary": "Download a full ZIP backup of the site",
        "operationId": "backupSite",
        "responses": { "200": { "description": "ZIP backup returned." } }
      }
    },

    "/logs": {
      "get": {
        "summary": "Retrieve server logs",
        "operationId": "getLogs",
        "responses": { "200": { "description": "Logs returned." } }
      }
    }
  }
}